+++ babel.config.cjs
@@ -10,4 +10,7 @@ module.exports = {
     ],
     ["@babel/preset-react", { runtime: "automatic" }],
   ],
+  plugins: [
+    "babel-plugin-transform-import-meta",
+  ],
 };